requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.7
django-ckeditor==6.5.1
django-decouple==2.1
django-etc==1.4.0
django-filter==23.1
django-formtools==2.4
django-hitcount==1.3.5
django-js-asset==2.0.0
django-otp==1.1.6
django-phonenumber-field==6.4.0
django-taggit==3.1.0
django-two-factor-auth==1.15.1
django-vote==2.4.0
filelock==3.9.0
idna==3.4
paystack==1.5.0
phonenumbers==8.13.9
Pillow==9.4.0
platformdirs==3.0.0
psycopg2==2.9.5
pyotp==2.8.0
pypng==0.20220715.0
python-decouple==3.8
qrcode==7.4.2
requests==2.28.2
simplejson==3.19.1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.15
virtualenv==20.20.0
virtualenvwrapper-win==1.2.7
whitenoise==6.4.0
